+++ app/src/main/AndroidManifest.xml
@@ -13,11 +13,14 @@
         android:theme="@style/Theme.RecipeGenius"
         tools:targetApi="31">
         <activity
+            android:name=".MainActivity2"
+            android:exported="false" />
+        <activity
             android:name=".MainActivity"
             android:exported="true"
-            android:label="@string/app_name">
-        </activity>
-        <activity android:name=".ui.questionnaire.QuestionnaireActivity"
+            android:label="@string/app_name"></activity>
+        <activity
+            android:name=".ui.questionnaire.QuestionnaireActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -26,4 +29,5 @@
             </intent-filter>
         </activity>
     </application>
+
 </manifest>
