+++ app/src/main/java/com/example/recipegenius/MainActivity.java
@@ -33,14 +33,14 @@ import java.time.YearMonth;
 import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 
-public class MainActivity extends AppCompatActivity implements CalendarAdapter.OnItemListener {
+public class MainActivity extends AppCompatActivity {
 
     private ActivityMainBinding binding;
     private DBHandler dbHandler;
 
-    private TextView monthYearText;
-    private RecyclerView calendarRecyclerView;
-    private LocalDate selectedDate;
+//    private TextView monthYearText;
+//    private RecyclerView calendarRecyclerView;
+//    private LocalDate selectedDate;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -65,10 +65,10 @@ public class MainActivity extends AppCompatActivity implements CalendarAdapter.O
         SharedPreferences dietFilters = context.getSharedPreferences("dietFilters", Context.MODE_PRIVATE);
         SharedPreferences allergyFilters = context.getSharedPreferences("allergyFilters", Context.MODE_PRIVATE);
 
-        setContentView(R.layout.fragment_calendar);
-        initWidgets();
-        selectedDate = LocalDate.now();
-        setMonthView();
+//        setContentView(R.layout.fragment_calendar);
+//        initWidgets();
+//        selectedDate = LocalDate.now();
+//        setMonthView();
 
 //        dbHandler = new DBHandler(MainActivity.this);
 //        Cursor cursor = dbHandler.getReadableDatabase().query("dietFilters", new String[] {"name"},  null, null, null, null, null, "100");
@@ -95,62 +95,62 @@ public class MainActivity extends AppCompatActivity implements CalendarAdapter.O
     }
 
 
-    private void initWidgets() {
-        calendarRecyclerView = findViewById(R.id.calendarRecyclerView);
-        monthYearText = findViewById(R.id.monthYearTV);
-
-    }
-
-    private void setMonthView() {
-        monthYearText.setText(monthYearFromDate(selectedDate));
-        ArrayList<String> daysInMonth = daysInMonthArray(selectedDate);
-
-        CalendarAdapter calendarAdapter = new CalendarAdapter(daysInMonth,this);
-        RecyclerView.LayoutManager layoutManager = new GridLayoutManager(getApplicationContext(),7);
-        calendarRecyclerView.setLayoutManager(layoutManager);
-        calendarRecyclerView.setAdapter(calendarAdapter);
-    }
-
-    private ArrayList<String> daysInMonthArray(LocalDate date) {
-        ArrayList<String> daysInMonthArray = new ArrayList<>();
-        YearMonth yearMonth = YearMonth.from(date);
-
-        int daysInMonth = yearMonth.lengthOfMonth();
-
-        LocalDate firstOfMonth = selectedDate.withDayOfMonth(1);
-        int dayOfWeek = firstOfMonth.getDayOfWeek().getValue();
-
-        for(int i =1; i<=42; i++){
-
-            if(i<=dayOfWeek || i>daysInMonth+dayOfWeek){
-                daysInMonthArray.add("");
-            }
-            else{
-                daysInMonthArray.add(String.valueOf(i-dayOfWeek));
-            }
-
-        }
-
-        return daysInMonthArray;
-
-
-    }
-
-    private String monthYearFromDate(LocalDate date){
-        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("MMMM yyyy");
-        return date.format(formatter);
-    }
-
-    public void previousMonthAction(View view){
-        selectedDate = selectedDate.minusMonths(1);
-        setMonthView();
-
-    }
-
-    public void nextMonthAction(View view){
-        selectedDate = selectedDate.plusMonths(1);
-        setMonthView();
-    }
+//    private void initWidgets() {
+//        calendarRecyclerView = findViewById(R.id.calendarRecyclerView);
+//        monthYearText = findViewById(R.id.monthYearTV);
+//
+//    }
+//
+//    private void setMonthView() {
+//        monthYearText.setText(monthYearFromDate(selectedDate));
+//        ArrayList<String> daysInMonth = daysInMonthArray(selectedDate);
+//
+//        CalendarAdapter calendarAdapter = new CalendarAdapter(daysInMonth,this);
+//        RecyclerView.LayoutManager layoutManager = new GridLayoutManager(getApplicationContext(),7);
+//        calendarRecyclerView.setLayoutManager(layoutManager);
+//        calendarRecyclerView.setAdapter(calendarAdapter);
+//    }
+//
+//    private ArrayList<String> daysInMonthArray(LocalDate date) {
+//        ArrayList<String> daysInMonthArray = new ArrayList<>();
+//        YearMonth yearMonth = YearMonth.from(date);
+//
+//        int daysInMonth = yearMonth.lengthOfMonth();
+//
+//        LocalDate firstOfMonth = selectedDate.withDayOfMonth(1);
+//        int dayOfWeek = firstOfMonth.getDayOfWeek().getValue();
+//
+//        for(int i =1; i<=42; i++){
+//
+//            if(i<=dayOfWeek || i>daysInMonth+dayOfWeek){
+//                daysInMonthArray.add("");
+//            }
+//            else{
+//                daysInMonthArray.add(String.valueOf(i-dayOfWeek));
+//            }
+//
+//        }
+//
+//        return daysInMonthArray;
+//
+//
+//    }
+//
+//    private String monthYearFromDate(LocalDate date){
+//        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("MMMM yyyy");
+//        return date.format(formatter);
+//    }
+//
+//    public void previousMonthAction(View view){
+//        selectedDate = selectedDate.minusMonths(1);
+//        setMonthView();
+//
+//    }
+//
+//    public void nextMonthAction(View view){
+//        selectedDate = selectedDate.plusMonths(1);
+//        setMonthView();
+//    }
 
 
     @Override
@@ -158,14 +158,14 @@ public class MainActivity extends AppCompatActivity implements CalendarAdapter.O
         NavController navController = Navigation.findNavController(this, R.id.nav_host_fragment_activity_main);
         return navController.navigateUp();
     }
-
-    @Override
-    public void onItemClick(int position, String dayText) {
-
-        if(dayText.equals("")){
-            String message = "Selected Data" + dayText + " " + monthYearFromDate(selectedDate);
-            Toast.makeText(this,message,Toast.LENGTH_LONG).show();
-        }
-
-    }
+//
+//    @Override
+//    public void onItemClick(int position, String dayText) {
+//
+//        if(dayText.equals("")){
+//            String message = "Selected Data" + dayText + " " + monthYearFromDate(selectedDate);
+//            Toast.makeText(this,message,Toast.LENGTH_LONG).show();
+//        }
+//
+//    }
 }
