+++ app/src/main/java/com/example/recipegenius/ui/calendar/CalendarFragment.java
@@ -1,5 +1,6 @@
 package com.example.recipegenius.ui.calendar;
 
+import android.content.Intent;
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -12,6 +13,7 @@ import androidx.fragment.app.Fragment;
 import androidx.lifecycle.ViewModelProvider;
 import androidx.navigation.Navigation;
 
+import com.example.recipegenius.MainActivity2;
 import com.example.recipegenius.R;
 import com.example.recipegenius.databinding.FragmentCalendarBinding;
 
@@ -22,6 +24,9 @@ public class CalendarFragment extends Fragment {
 
     public View onCreateView(@NonNull LayoutInflater inflater,
                              ViewGroup container, Bundle savedInstanceState) {
+
+        Intent i = new Intent(getActivity(),MainActivity2.class);
+        startActivity(i);
         CalendarViewModel homeViewModel =
                 new ViewModelProvider(this).get(CalendarViewModel.class);
